[foilnotes] Fixed current page detection
Something has changed in the Silica Page implementation that broke the
previous method of detecting the current top level page (encrypted vs
plaintext). Which in turn broke the cover behavior.

Replaced it with (hopefully) a more reliable method. Basically, the
state of the encrypted is always opposite to the state of the plaintext
(attached) page which can be determined reliably.